refactor(pwsh): eliminating templating in setup including windows setup in favor of powershell automation
Templating is is confusing, and problematic in terms of code editing
since formatters, lsps, and linters do not support it.
So it is sensible to minimalise templating and using either shell
methods or the applications inbuilt configuration / api to resolve
config differences per os.
Of course non-templating solutions are only preferable if they do not
have a significant overhead, since templating runs once and hence
faster.