feat(workspacer): Adding intsall path list testing for the templating of the prerpocessor
In order to give proper '#r library.dll' directives we have to know the
Workspacer install directory. For this purpose, I introduced a list of
install directories that we check for the workspacer.exe. If a directory
has an executable then we assign it as the install path. I favor scoop
over winget and locality over globality, hence the end or the list is
the scoop user installation of Workspace, i.e. it is the highest
priority.