feat(LazyVim): try out setting filetype to C for OpenCL files, add C-like language settings
It turned out that much of the components of the Clangd LazyVim extra
supports OpenCL natively, hence they can be used for OpenCL without much
configuration, making my user-level opencl extra redundant. The
vim-opencl syntax highlighter plugin is especially poor.

On the other hand OpenCl is basically C99 without malloc, and recursion,
so it is not a stretch.

In addition I created an autocommand for C-like language file settings.
Currently it contains only the 8 character wide tab indentation
instructions.