feat(Windows): Replacing lf to yazi, use git bundled unix tools instead of the others
After a more careful look at the Git For Windows package, I found out
that much of the utilities, I scooped down are already here. To
eliminate this redundancy, I uninstalled the scooped tools and shimmed
the tools in the Git For Windows (scoop) installation directory.
This only made my config a slightly more lightweight. (Anaconda make it
very cumbersome).